requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
ruff>=0.6.0
python-dotenv>=1.0.0
